pyproject.toml
[project]
name = "overwatchstatsmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "mcp[cli]>=1.26.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "python-dotenv>=1.0",
    "pillow>=12.1.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "testcontainers>=4.0",
    "psycopg2-binary>=2.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["src"]
